+++ src/api/namespaces/atomicmarket/routes/auctions.ts
@@ -4,7 +4,6 @@ import { AtomicMarketNamespace, AuctionApiState } from '../index';
 import { HTTPServer } from '../../../server';
 import { formatAuction } from '../format';
 import { fillAuctions } from '../filler';
-import { buildAuctionFilter, hasListingFilter } from '../utils';
 import {
     actionGreylistParameters,
     dateBoundaryParameters,
@@ -13,150 +12,25 @@ import {
     primaryBoundaryParameters
 } from '../../../docs';
 import { extendedAssetFilterParameters, atomicDataFilter, baseAssetFilterParameters } from '../../atomicassets/openapi';
-import { buildBoundaryFilter, filterQueryArgs } from '../../utils';
 import { listingFilterParameters } from '../openapi';
-import { buildGreylistFilter, hasAssetFilter, hasDataFilters } from '../../atomicassets/utils';
 import {
-    applyActionGreylistFilters,
     createSocketApiNamespace,
     extractNotificationIdentifiers,
-    getContractActionLogs, respondApiError
 } from '../../../utils';
 import ApiNotificationReceiver from '../../../notification';
 import { NotificationData } from '../../../../filler/notifier';
 import { eosioTimestampToDate } from '../../../../utils/eosio';
-import QueryBuilder from '../../../builder';
+import { getAuctionAction, getAuctionLogsAction, getAuctionsAction, getAuctionsCountAction } from './handlers/auctions';
 
 export function auctionsEndpoints(core: AtomicMarketNamespace, server: HTTPServer, router: express.Router): any {
-    router.all(['/v1/auctions', '/v1/auctions/_count'], server.web.caching(), async (req, res) => {
-        try {
-            const args = filterQueryArgs(req, {
-                page: {type: 'int', min: 1, default: 1},
-                limit: {type: 'int', min: 1, max: 100, default: 100},
-                sort: {
-                    type: 'string',
-                    values: [
-                        'created', 'updated', 'ending', 'auction_id', 'price',
-                        'template_mint'
-                    ],
-                    default: 'created'
-                },
-                order: {type: 'string', values: ['asc', 'desc'], default: 'desc'}
-            });
+    const {caching, returnAsJSON} = server.web;
 
-            const query = new QueryBuilder(
-                'SELECT listing.auction_id ' +
-                'FROM atomicmarket_auctions listing ' +
-                'JOIN atomicmarket_tokens "token" ON (listing.market_contract = "token".market_contract AND listing.token_symbol = "token".token_symbol)'
-            );
+    router.all('/v1/auctions', caching(), returnAsJSON(getAuctionsAction, core));
+    router.all('/v1/auctions/_count', caching(), returnAsJSON(getAuctionsCountAction, core));
 
-            query.equal('listing.market_contract', core.args.atomicmarket_account);
+    router.all('/v1/auctions/:auction_id', caching(), returnAsJSON(getAuctionAction, core));
 
-            query.addCondition(
-                'NOT EXISTS (' +
-                'SELECT * FROM atomicmarket_auctions_assets auction_asset ' +
-                'WHERE auction_asset.market_contract = listing.market_contract AND auction_asset.auction_id = listing.auction_id AND ' +
-                'NOT EXISTS (SELECT * FROM atomicassets_assets asset WHERE asset.contract = auction_asset.assets_contract AND asset.asset_id = auction_asset.asset_id)' +
-                ')'
-            );
-
-            buildAuctionFilter(req, query);
-            buildGreylistFilter(req, query, {collectionName: 'listing.collection_name'});
-            buildBoundaryFilter(
-                req, query, 'listing.auction_id', 'int',
-                args.sort === 'updated' ? 'listing.updated_at_time' : 'listing.created_at_time'
-            );
-
-            if (req.originalUrl.search('/_count') >= 0) {
-                const countQuery = await server.query(
-                    'SELECT COUNT(*) counter FROM (' + query.buildString() + ') x',
-                    query.buildValues()
-                );
-
-                return res.json({success: true, data: countQuery.rows[0].counter, query_time: Date.now()});
-            }
-
-            const sortMapping: {[key: string]: {column: string, nullable: boolean, numericIndex: boolean}} = {
-                auction_id: {column: 'listing.auction_id', nullable: false, numericIndex: true},
-                ending: {column: 'listing.end_time', nullable: false, numericIndex: true},
-                created: {column: 'listing.created_at_time', nullable: false, numericIndex: true},
-                updated: {column: 'listing.updated_at_time', nullable: false, numericIndex: true},
-                price: {column: 'listing.price', nullable: true, numericIndex: false},
-                template_mint: {column: 'LOWER(listing.template_mint)', nullable: true, numericIndex: false}
-            };
-
-            const ignoreIndex = (hasAssetFilter(req) || hasDataFilters(req) || hasListingFilter(req)) && sortMapping[args.sort].numericIndex;
-
-            query.append('ORDER BY ' + sortMapping[args.sort].column + (ignoreIndex ? ' + 1 ' : ' ') + args.order + ' ' + (sortMapping[args.sort].nullable ? 'NULLS LAST' : '') + ', listing.auction_id ASC');
-            query.paginate(args.page, args.limit);
-
-            const auctionResult = await server.query(query.buildString(), query.buildValues());
-
-            const auctionLookup: {[key: string]: any} = {};
-            const result = await server.query(
-                'SELECT * FROM atomicmarket_auctions_master WHERE market_contract = $1 AND auction_id = ANY ($2)',
-                [core.args.atomicmarket_account, auctionResult.rows.map(row => row.auction_id)]
-            );
-
-            result.rows.reduce((prev, current) => {
-                prev[String(current.auction_id)] = current;
-
-                return prev;
-            }, auctionLookup);
-
-            const auctions = await fillAuctions(
-                server, core.args.atomicassets_account,
-                auctionResult.rows.map((row) => formatAuction(auctionLookup[String(row.auction_id)]))
-            );
-
-            res.json({success: true, data: auctions, query_time: Date.now()});
-        } catch (error) {
-            return respondApiError(res, error);
-        }
-    });
-
-    router.all('/v1/auctions/:auction_id', server.web.caching(), async (req, res) => {
-        try {
-            const query = await server.query(
-                'SELECT * FROM atomicmarket_auctions_master WHERE market_contract = $1 AND auction_id = $2',
-                [core.args.atomicmarket_account, req.params.auction_id]
-            );
-
-            if (query.rowCount === 0) {
-                res.status(416).json({success: false, message: 'Auction not found'});
-            } else {
-                const auctions = await fillAuctions(
-                    server, core.args.atomicassets_account, query.rows.map((row) => formatAuction(row))
-                );
-
-                res.json({success: true, data: auctions[0], query_time: Date.now()});
-            }
-        } catch (error) {
-            return respondApiError(res, error);
-        }
-    });
-
-    router.all('/v1/auctions/:auction_id/logs', server.web.caching(), (async (req, res) => {
-        const args = filterQueryArgs(req, {
-            page: {type: 'int', min: 1, default: 1},
-            limit: {type: 'int', min: 1, max: 100, default: 100},
-            order: {type: 'string', values: ['asc', 'desc'], default: 'asc'}
-        });
-
-        try {
-            res.json({
-                success: true,
-                data: await getContractActionLogs(
-                    server, core.args.atomicmarket_account,
-                    applyActionGreylistFilters(['lognewauct', 'logauctstart', 'cancelauct', 'auctclaimbuy', 'auctclaimsel'], args),
-                    {auction_id: req.params.auction_id},
-                    (args.page - 1) * args.limit, args.limit, args.order
-                ), query_time: Date.now()
-            });
-        } catch (error) {
-            return respondApiError(res, error);
-        }
-    }));
+    router.all('/v1/auctions/:auction_id/logs', server.web.caching(), returnAsJSON(getAuctionLogsAction, core));
 
     return {
         tag: {
